app/requirements.txt
uwsgi 
flask
flask-cors
Flask-Sessions
